requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.1
asn1crypto==1.5.1
Babel==2.10.3
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-admin-rangefilter==0.8.4
django-easy-pdf==0.1.1
django-filter==21.1
django-import-export==2.8.0
django-jazzmin==2.5.0
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.33.3
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
import-export==0.2.67.dev6
jmespath==1.0.0
lxml==4.9.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
oscrypto==1.3.0
Pillow==9.1.1
pipreqs==0.4.11
pycparser==2.21
pydyf==0.2.0
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyMySQL==1.0.2
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportbro-fpdf==1.7.11
reportbro-lib==2.1.1
reportbro-simpleeval==0.9.11
reportlab==3.6.10
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
weasyprint==55.0
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
yarg==0.1.9
zopfli==0.2.1
